Moved all volume subcommands under volume.go
This is to allow for other commands to share the same subcommands (see:
volume vs screen both needing inc, etc...
